pedestrian-box/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pedestrian_box"
version = "0.1.0"
description = "Pedestrian detection project"
authors = [
    { name = "Anatolii" }
]
requires-python = ">=3.10"

dependencies = [
    "torch>=2.0",
    "torchvision",
    "hydra-core",
    "mlflow",
    "pandas",
    "numpy",
    "matplotlib",
    "torchmetrics",
    "kagglehub",
    "pillow",
    "pyyaml",
    "streamlit",
    "ultralytics"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "black",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
